fix: comment, tag 앱 없애기
post로 다 옮김. 근데 로직은 그대로라 수정필요
