--- awesome_blog/comment/urls.py
@@ -1,11 +1,0 @@
-from django.urls import path, include
-from rest_framework.routers import DefaultRouter
-
-from .views import CommentViewSet
-
-router = DefaultRouter()
-router.register(r'comments', CommentViewSet)
-
-urlpatterns = [
-    path('', include(router.urls)),
-]
