+++ awesome_blog/post/admin.py
@@ -1,5 +1,5 @@
 from django.contrib import admin
-from .models import Post
+from .models import Post, Comment, Tag
 
 
 @admin.register(Post)
@@ -15,3 +15,35 @@ class PostAdmin(admin.ModelAdmin):
         return ', '.join([tag.content for tag in obj.tags.all()])
 
     display_tags.short_description = 'Tags'
+
+
+@admin.register(Comment)
+class CommentAdmin(admin.ModelAdmin):
+    list_display = ('post', 'created_by', 'created_at', 'updated_at', 'display_tags')
+    search_fields = ('post', 'created_by', 'tags')
+    list_filter = ('created_at', 'updated_at', 'tags', 'created_by')
+    date_hierarchy = 'created_at'
+    ordering = ('-created_at',)
+
+    def display_tags(self, obj):
+        """Create a string for the Tags. This is required to display tags in List display."""
+        return ', '.join([tag.content for tag in obj.tags.all()])
+
+    display_tags.short_description = 'Tags'
+
+
+class PostInline(admin.TabularInline):
+    model = Post.tags.through
+    extra = 0
+
+
+class CommentInline(admin.TabularInline):
+    model = Comment.tags.through
+    extra = 0
+
+
+@admin.register(Tag)
+class TagAdmin(admin.ModelAdmin):
+    list_display = ('content',)
+    inlines = [PostInline, CommentInline]
+    search_fields = ('content',)
