+++ awesome_blog/post/management/commands/create_fake_comments.py
