--- awesome_blog/tag/serializers.py
@@ -1,15 +1,0 @@
-from rest_framework import serializers
-from rest_framework.fields import empty
-
-from .models import Tag
-
-
-class TagSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Tag
-        fields = ['content']
-
-    def create(self, validated_data):
-        tag, created = Tag.objects.get_or_create(content=validated_data['content'])
-        return tag
-
